arch-fix: extract cross-rule helpers into fencepos and tablefmt (plan/154)
Lift the helpers that four rules borrowed from fencedcodestyle
(FenceCharAt, FenceOpenLine, FenceOpenLineRange, FenceCloseLine,
FenceCloseLineRange) into a sibling internal/rules/fencepos package
exporting CharAt, OpenLine, OpenLineRange, CloseLine, CloseLineRange.
Likewise, lift tableformat.FormatString plus its parsing/formatting
machinery into internal/rules/tablefmt; the donor consumes the new
Violations and FormatLines surfaces. No rule imports another rule.

TestRulesDoNotImportEachOther in internal/integration/ guards the
boundary going forward by parsing every non-test Go file under
internal/rules/ and failing on imports of another rule package
outside the documented helper allowlist (astutil, settings,
fencepos, tablefmt). The blank-import barrel internal/rules/all/
is exempt.
