+++ internal/rules/orderedlistnumbering/rule.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/jeduden/mdsmith/internal/lint"
 	"github.com/jeduden/mdsmith/internal/rule"
-	"github.com/jeduden/mdsmith/internal/rules/fencedcodestyle"
+	"github.com/jeduden/mdsmith/internal/rules/fencepos"
 	"github.com/jeduden/mdsmith/internal/rules/settings"
 	"github.com/yuin/goldmark/ast"
 )
@@ -386,7 +386,7 @@ func blockLastLine(f *lint.File, n ast.Node) int {
 	// when a marker digit-width changes (otherwise the fence outdents
 	// and breaks the block).
 	if fcb, ok := n.(*ast.FencedCodeBlock); ok {
-		return fencedcodestyle.FenceCloseLine(f, fcb)
+		return fencepos.CloseLine(f, fcb)
 	}
 	if n.Lines().Len() > 0 {
 		return f.LineOfOffset(n.Lines().At(n.Lines().Len() - 1).Start)
